Move locateSolrHome to new SolrPaths Move "User Files" stuff to SolrPaths Inline constant project field
